Add files via upload
Twitter Sentiment Analysis determines whether a tweet is positive, negative, or neutral emotions within the text of a tweet using NLP and ML models. Sentiment analysis or opinion mining refers to identifying as well as classifying the sentiments that are expressed in the text source. Natural language processing (NLP) is a branch of artificial intelligence (AI) that enables computers to comprehend, generate, and manipulate human language concerned with giving computers the ability to understand text and spoken words in much the same way human beings can.